[Pro Feat] 유물, 유물효과, 유물획득 시스템 구현 및 유물 데이터 베이스 구현
+) 레벨 데이터 베이스 수정 (0~100 --> 0~1)

+++ SpaceAttack/Assets/Scripts/Database/Level/LevelData.cs
@@ -9,8 +9,8 @@ public class LevelData
     public string levelKey;
     public int level;
     public int maxEX;
-    public int? damageCorrection;
+    public float? damageCorrection;
     public int? heartCorrection;
-    public int? speedCorrection;
+    public float? speedCorrection;
     public int? unlockedSkill;
 }
