+++ SpaceAttack/Assets/Scripts/Database/Level/LevelSO.cs
@@ -7,9 +7,9 @@ public class LevelSO : ScriptableObject
     public string levelKey;
     public int level;
     public int maxEX;
-    public int damageCorrection;
+    public float damageCorrection;
     public int heartCorrection;
-    public int speedCorrection;
+    public float speedCorrection;
     public int unlockedSkill;
     public bool unlockability;
     public bool correctability;
